+++ src/resolve/metadataResolver.ts
@@ -179,8 +179,8 @@ export class MetadataResolver {
           // any of the following 3 options is considered a good match
           // mixedContent and bundles don't have a suffix to match
           ['mixedContent', 'bundle'].includes(type.strategies?.adapter) ||
-          // the suffix matches the type we think it is
-          (type.suffix && fsPath.endsWith(`${type.suffix}${META_XML_SUFFIX}`)) ||
+          // the file suffix (in source or mdapi format) matches the type suffix we think it is
+          (type.suffix && [type.suffix, `${type.suffix}${META_XML_SUFFIX}`].some((s) => fsPath.endsWith(s))) ||
           // the type has children and the path also includes THAT directory
           (type.children?.types &&
             Object.values(type.children?.types)
@@ -235,10 +235,12 @@ export class MetadataResolver {
       const { directoryName, inFolder } = type;
       const parts = dirPath.split(sep);
       const folderOffset = inFolder ? 2 : 1;
-      const typeDirectoryIndex = parts.indexOf(directoryName);
+      const typeDirectoryIndex = parts.lastIndexOf(directoryName);
       if (
         typeDirectoryIndex === -1 ||
         parts.length - folderOffset <= typeDirectoryIndex ||
+        // ex: /lwc/folder/lwc/cmp
+        this.tree.readDirectory(dirPath).includes(type.directoryName) ||
         // types with children may want to resolve them individually
         type.children
       ) {