+++ test/resolve/adapters/bundleSourceAdapter.test.ts
@@ -6,12 +6,13 @@
  */
 
 import { expect } from 'chai';
-import { bundle } from '../../mock';
+import { bundle, lwcBundle } from '../../mock';
 import { BundleSourceAdapter } from '../../../src/resolve/adapters';
 import { CONTENT_PATH } from '../../mock/type-constants/auraBundleConstant';
+import { CONTENT_PATH as LWC_CONTENT_PATH } from '../../mock/type-constants/lwcBundleConstant';
 import { RegistryAccess } from '../../../src';
 
-describe('BundleSourceAdapter', () => {
+describe('BundleSourceAdapter with AuraBundle', () => {
   const registryAccess = new RegistryAccess();
   const adapter = new BundleSourceAdapter(bundle.COMPONENT.type, registryAccess, undefined, bundle.COMPONENT.tree);
 
@@ -36,5 +37,36 @@ describe('BundleSourceAdapter', () => {
   it('Should return expected SourceComponent when given a source path', () => {
     const randomSource = bundle.SOURCE_PATHS[1];
     expect(adapter.getComponent(randomSource)).to.deep.equal(bundle.COMPONENT);
+  });
+
+  describe('deeply nested LWC', () => {
+    const lwcAdapter = new BundleSourceAdapter(
+      lwcBundle.COMPONENT.type,
+      registryAccess,
+      undefined,
+      lwcBundle.COMPONENT.tree
+    );
+    it('Should return expected SourceComponent when given a root metadata xml path', () => {
+      expect(lwcAdapter.getComponent(lwcBundle.XML_PATH)).to.deep.equal(lwcBundle.COMPONENT);
+    });
+
+    it('Should return expected SourceComponent when given a lwcBundle directory', () => {
+      expect(lwcAdapter.getComponent(lwcBundle.CONTENT_PATH)).to.deep.equal(lwcBundle.COMPONENT);
+    });
+
+    it('Should return expected SourceComponent when given a source path', () => {
+      const randomSource = lwcBundle.SOURCE_PATHS[1];
+      expect(lwcAdapter.getComponent(randomSource)).to.deep.equal(lwcBundle.COMPONENT);
+    });
+
+    it('Should exclude nested empty bundle directories', () => {
+      const emptyBundleAdapter = new BundleSourceAdapter(
+        lwcBundle.EMPTY_BUNDLE.type,
+        registryAccess,
+        undefined,
+        lwcBundle.EMPTY_BUNDLE.tree
+      );
+      expect(emptyBundleAdapter.getComponent(LWC_CONTENT_PATH)).to.be.undefined;
+    });
   });
 });